feat(nms): Add subscriber bulk update
