+++ nms/packages/magmalte/app/views/subscriber/SubscriberAddDialog.js
@@ -14,31 +14,22 @@
  * @format
  */
 import type {ActionQuery} from '../../components/ActionTable';
-import type {SubscriberActionType} from './SubscriberUtils';
-import type {
-  core_network_types,
-  subscriber,
-} from '../../../generated/MagmaAPIBindings';
+import type {EditProps} from './SubscriberEditDialog';
+import type {SubscriberActionType, SubscriberInfo} from './SubscriberUtils';
 
 import ActionTable from '../../components/ActionTable';
 import Alert from '@material-ui/lab/Alert';
 import AlertTitle from '@material-ui/lab/AlertTitle';
 import ApnContext from '../../components/context/ApnContext';
 import Button from '@material-ui/core/Button';
-import CardTitleRow from '../../components/layout/CardTitleRow';
 import Checkbox from '@material-ui/core/Checkbox';
 import CloudUploadIcon from '@material-ui/icons/CloudUpload';
 import Dialog from '@material-ui/core/Dialog';
 import DialogActions from '@material-ui/core/DialogActions';
 import DialogContent from '@material-ui/core/DialogContent';
 import DialogTitle from '../../theme/design-system/DialogTitle';
-import EditSubscriberApnStaticIps from './SubscriberApnStaticIpsEdit';
-import EditSubscriberTrafficPolicy from './SubscriberTrafficPolicyEdit';
 import FormControl from '@material-ui/core/FormControl';
-import FormLabel from '@material-ui/core/FormLabel';
 import Grid from '@material-ui/core/Grid';
-import Link from '@material-ui/core/Link';
-import List from '@material-ui/core/List';
 import ListItemText from '@material-ui/core/ListItemText';
 import LteNetworkContext from '../../components/context/LteNetworkContext';
 import MenuItem from '@material-ui/core/MenuItem';
@@ -47,66 +38,29 @@ import PolicyContext from '../../components/context/PolicyContext';
 import React from 'react';
 import Select from '@material-ui/core/Select';
 import SubscriberContext from '../../components/context/SubscriberContext';
-import Tab from '@material-ui/core/Tab';
-import Tabs from '@material-ui/core/Tabs';
 import Text from '../../theme/design-system/Text';
 import Tooltip from '@material-ui/core/Tooltip';
-import TypedSelect from '@fbcnms/ui/components/TypedSelect';
 import nullthrows from '@fbcnms/util/nullthrows';
 
-import {AltFormField, PasswordInput} from '../../components/FormField';
-import {CoreNetworkTypes, SUBSCRIBER_ADD_ERRORS} from './SubscriberUtils';
-import {DropzoneArea} from 'material-ui-dropzone';
+import {CoreNetworkTypes, validateSubscribers} from './SubscriberUtils';
+import {PasswordInput} from '../../components/FormField';
 import {SelectEditComponent} from '../../components/ActionTable';
-import {base64ToHex, hexToBase64, isValidHex} from '@fbcnms/util/strings';
-import {colors, typography} from '../../theme/default';
+import {SubscriberDetailsUpload} from './SubscriberUpload';
+import {colors} from '../../theme/default';
 import {forwardRef} from 'react';
 import {handleSubscriberQuery} from '../../state/lte/SubscriberState';
 import {makeStyles} from '@material-ui/styles';
-import {useContext, useEffect, useState} from 'react';
-import {useEnqueueSnackbar} from '@fbcnms/ui/hooks/useSnackbar';
+import {useContext, useState} from 'react';
 import {useRouter} from '@fbcnms/ui/hooks';
 
-const useStyles = makeStyles(theme => ({
-  alert: {
-    backgroundColor: colors.primary.white,
-    padding: '20px 40px 20px 40px',
-  },
-  dashboardRoot: {
-    margin: theme.spacing(3),
-    flexGrow: 1,
+const useStyles = makeStyles(() => ({
+  dialogTitle: {
+    textTransform: 'capitalize',
+    backgroundColor: colors.primary.brightGray,
   },
   tabBar: {
     backgroundColor: colors.primary.brightGray,
     color: colors.primary.white,
-  },
-  tabs: {
-    color: colors.primary.white,
-  },
-  tab: {
-    fontSize: '18px',
-    textTransform: 'none',
-  },
-  tabLabel: {
-    padding: '16px 0 16px 0',
-    display: 'flex',
-    alignItems: 'center',
-  },
-  tabIconLabel: {
-    marginRight: '8px',
-  },
-  appBarBtn: {
-    color: colors.primary.white,
-    background: colors.primary.comet,
-    fontFamily: typography.button.fontFamily,
-    fontWeight: typography.button.fontWeight,
-    fontSize: typography.button.fontSize,
-    lineHeight: typography.button.lineHeight,
-    letterSpacing: typography.button.letterSpacing,
-
-    '&:hover': {
-      background: colors.primary.mirage,
-    },
   },
   input: {
     display: 'inline-flex',
@@ -138,150 +92,15 @@ const useStyles = makeStyles(theme => ({
   },
 }));
 
-const MAX_UPLOAD_FILE_SZ_BYTES = 10 * 1024 * 1024;
-const SUBSCRIBER_TITLE = 'Subscriber';
-const TRAFFIC_TITLE = 'Traffic Policy';
-const STATIC_IPS_TITLE = 'APNs Static IPs';
-const UPLOAD_DOC_LINK =
-  'https://docs.magmacore.org/docs/nms/subscriber#uploading-a-subscriber-csv-file';
-const ADD_INSTRUCTIONS =
-  'You can download this template that automatically maps the fields. Find more instruction in ';
-const DELETE_INSTRUCTIONS =
-  'You can export all subscribers and select the subscribers you want to delete. Find more instruction in ';
-const EDIT_INSTRUCTIONS =
-  'You can export all subscribers to edit and upload the file. Find more instruction in ';
-
-export type SubscriberInfo = {
-  name: string,
-  imsi: string,
-  authKey: string,
-  authOpc: string,
-  state: 'INACTIVE' | 'ACTIVE',
-  forbiddenNetworkTypes: core_network_types,
-  dataPlan: string,
-  apns: Array<string>,
-  policies?: Array<string>,
-};
-
-const SUB_NAME_OFFSET = 0;
-const SUB_IMSI_OFFSET = 1;
-const SUB_AUTH_KEY_OFFSET = 2;
-const SUB_AUTH_OPC_OFFSET = 3;
-const SUB_STATE_OFFSET = 4;
-const SUB_FORBIDDEN_NETWORK_TYPE_OFFSET = 5;
-const SUB_DATAPLAN_OFFSET = 6;
-const SUB_APN_OFFSET = 7;
-const SUB_POLICY_OFFSET = 8;
-const SUB_MAX_FIELDS = 9;
 const forbiddenNetworkTypes = Object.keys(CoreNetworkTypes).map(
   key => CoreNetworkTypes[key],
 );
-
-function parseSubscriber(line: string) {
-  const items = line.split(',').map(item => item.trim());
-  if (items.length > SUB_MAX_FIELDS) {
-    throw new Error(
-      `Too many fields to parse, expected ${SUB_MAX_FIELDS} fields, received ${items.length} fields`,
-    );
-  }
-  return {
-    name: items[SUB_NAME_OFFSET],
-    imsi: items[SUB_IMSI_OFFSET],
-    authKey: items[SUB_AUTH_KEY_OFFSET],
-    authOpc: items[SUB_AUTH_OPC_OFFSET],
-    state: items[SUB_STATE_OFFSET] === 'ACTIVE' ? 'ACTIVE' : 'INACTIVE',
-    forbiddenNetworkTypes: forbiddenNetworkTypes.filter(value =>
-      items[SUB_FORBIDDEN_NETWORK_TYPE_OFFSET]?.split('|')
-        .map(item => item.trim())
-        .filter(Boolean)
-        .includes(value),
-    ),
-    dataPlan: items[SUB_DATAPLAN_OFFSET],
-    apns: items[SUB_APN_OFFSET]?.split('|')
-      .map(item => item.trim())
-      .filter(Boolean),
-    policies: items?.[SUB_POLICY_OFFSET]?.split('|')
-      .map(item => item.trim())
-      .filter(Boolean),
-  };
-}
-
-function parseSubscriberFile(fileObj: File) {
-  const reader = new FileReader();
-  const subscribers = [];
-  return new Promise((resolve, reject) => {
-    if (fileObj.size > MAX_UPLOAD_FILE_SZ_BYTES) {
-      reject(
-        'file size exceeds max upload size of 10MB, please upload smaller file',
-      );
-      return;
-    }
-
-    reader.onload = async e => {
-      try {
-        if (!(e.target instanceof FileReader)) {
-          reject('invalid target type');
-          return;
-        }
-
-        const text = e.target.result;
-        if (typeof text !== 'string') {
-          reject('invalid file content');
-          return;
-        }
-
-        for (const line of text
-          .split('\n')
-          .map(item => item.trim())
-          .filter(Boolean)) {
-          subscribers.push(parseSubscriber(line));
-        }
-      } catch (e) {
-        reject('Failed parsing the file ' + fileObj.name + '. ' + e?.message);
-        return;
-      }
-      resolve(subscribers);
-    };
-    reader.readAsText(fileObj);
-  });
-}
-
-export function EditSubscriberButton(props: EditProps) {
-  const [open, setOpen] = useState(false);
-  return (
-    <>
-      <SubscriberEditDialog
-        editProps={props}
-        open={open}
-        onClose={() => setOpen(false)}
-        subscriberAction={''}
-      />
-      <Button
-        component="button"
-        data-testid={props.editTable}
-        variant="text"
-        onClick={() => setOpen(true)}>
-        {'Edit'}
-      </Button>
-    </>
-  );
-}
-
-const EditTableType = {
-  subscriber: 0,
-  trafficPolicy: 1,
-  staticIps: 2,
-};
-
-type EditProps = {
-  editTable: $Keys<typeof EditTableType>,
-};
 
 type ActionDialogProps = {
   open: boolean,
   onClose: () => void,
   editProps?: EditProps,
-  onSave?: (
+  onSave: (
     subscribers: Array<SubscriberInfo>,
     selectedSubscribers?: Array<string>,
   ) => void,
@@ -289,21 +108,11 @@ type ActionDialogProps = {
   subscriberAction: SubscriberActionType,
 };
 
-type DialogProps = {
-  open: boolean,
-  onClose: () => void,
-  editProps?: EditProps,
-  onSave?: (
-    subscribers: Array<SubscriberInfo>,
-    selectedSubscribers?: Array<string>,
-  ) => void,
-  error?: string,
-};
-
 /**
  * Dialog used to Add/Delete/Update subscribers
  */
 export function AddSubscriberDialog(props: ActionDialogProps) {
+  const classes = useStyles();
   return (
     <>
       <Dialog
@@ -314,6 +123,7 @@ export function AddSubscriberDialog(props: ActionDialogProps) {
         }}
         maxWidth="xl">
         <DialogTitle
+          classes={{root: classes.dialogTitle}}
           onClose={props.onClose}
           label={`${props.subscriberAction} Subscriber(s)`}
         />
@@ -324,237 +134,43 @@ export function AddSubscriberDialog(props: ActionDialogProps) {
   );
 }
 
-export type EditSubscriberProps = {
-  subscriberState: subscriber,
-  onSubscriberChange: (key: string, val: string | number | {}) => void,
-  inputClass: string,
-  onTrafficPolicyChange: (
-    key: string,
-    val: string | number | {},
-    index: number,
-  ) => void,
-  onDeleteApn: (apn: {}) => void,
-  onAddApnStaticIP: () => void,
-  subProfiles: {},
-  subscriberStaticIPRows: Array<subscriberStaticIpsRowType>,
-  forbiddenNetworkTypes: Array<subscriberForbiddenNetworkTypes>,
-  authKey: string,
-  authOpc: string,
-  setAuthKey: (key: string) => void,
-  setAuthOpc: (key: string) => void,
-};
+/**
+ * Dialog content used to Add/Delete/Update subscribers
+ * Displays upload subscriber dropzone or subscriber table
+ */
+function SubscriberDetailsDialogContent(props: ActionDialogProps) {
+  const [addError, setAddError] = useState([]);
+  const [subscribers, setSubscribers] = useState<Array<SubscriberInfo>>([]);
+  const [upload, setUpload] = useState(false);
+  const [rowAdd, setRowAdd] = useState(false);
 
-export function SubscriberEditDialog(props: DialogProps) {
-  const {editProps} = props;
-  const enqueueSnackbar = useEnqueueSnackbar();
-  const [tabPos, setTabPos] = useState(
-    editProps ? EditTableType[editProps.editTable] : 0,
-  );
-  const ctx = useContext(SubscriberContext);
-  const lteCtx = useContext(LteNetworkContext);
-  const classes = useStyles();
-  const {match} = useRouter();
-  const subscriberId = nullthrows(match.params.subscriberId);
-  const [subscriberState, setSubscriberState] = useState<subscriber>(
-    ctx.state[subscriberId],
-  );
-
-  const [authKey, setAuthKey] = useState(
-    subscriberState.lte.auth_key
-      ? base64ToHex(subscriberState.lte.auth_key)
-      : '',
-  );
-  const [authOpc, setAuthOpc] = useState(
-    subscriberState.lte.auth_opc != null
-      ? base64ToHex(subscriberState.lte.auth_opc)
-      : '',
-  );
-  const [subscriberStaticIPRows, setSubscriberStaticIPRows] = useState<
-    Array<subscriberStaticIpsRowType>,
-  >(
-    Object.keys(ctx.state[subscriberId].config.static_ips || {}).map(
-      (apn: string) => {
-        return {
-          apnName: apn,
-          staticIp: ctx.state[subscriberId].config.static_ips?.[apn] || '',
-        };
-      },
-    ),
-  );
-
-  const subscriberCoreNetwork = Array<subscriberForbiddenNetworkTypes>(
-    Object.keys(CoreNetworkTypes).map((key: string) => {
-      return {
-        nwTypes: key,
-      };
-    }),
-  );
-
-  const [error, setError] = useState('');
-  useEffect(() => {
-    setTabPos(props.editProps ? EditTableType[props.editProps.editTable] : 0);
-  }, [props.editProps]);
-
-  const onClose = () => {
-    props.onClose();
-  };
-
-  // we are doing this to ensure we can map subprofiles from an array
-  // for e.g. ['foo', 'default'] -> {foo: 'foo', default: 'default}
-  // this is done because TypedSelect expects items in this form to verify
-  // if the passed in input is of expected type
-  const subProfiles = Array.from(
-    new Set(Object.keys(lteCtx.state.cellular?.epc?.sub_profiles || {})).add(
-      'default',
-    ),
-  ).reduce(function (o, v) {
-    o[v] = v;
-    return o;
-  }, {});
-
-  const subscriberProps: EditSubscriberProps = {
-    subscriberState: subscriberState,
-    onSubscriberChange: (key: string, val) => {
-      setSubscriberState({...subscriberState, [key]: val});
-    },
-    onTrafficPolicyChange: (key: string, val, index: number) => {
-      const rows = subscriberStaticIPRows;
-      rows[index][key] = val;
-      setSubscriberStaticIPRows([...rows]);
-    },
-    onDeleteApn: (apn: {}) => {
-      setSubscriberStaticIPRows([
-        ...subscriberStaticIPRows.filter(
-          (deletedApn: subscriberStaticIpsRowType) => apn !== deletedApn,
-        ),
-      ]);
-    },
-    onAddApnStaticIP: () => {
-      setSubscriberStaticIPRows([
-        ...subscriberStaticIPRows,
-        {apnName: '', staticIp: ''},
-      ]);
-    },
-    subProfiles: subProfiles,
-    subscriberStaticIPRows: subscriberStaticIPRows,
-    forbiddenNetworkTypes: subscriberCoreNetwork,
-    authKey: authKey,
-    authOpc: authOpc,
-    setAuthKey: (key: string) => setAuthKey(key),
-    setAuthOpc: (key: string) => setAuthOpc(key),
-    inputClass: classes.input,
-  };
-
-  const onSave = async () => {
-    try {
-      if (authOpc !== '') {
-        if (isValidHex(authOpc)) {
-          subscriberState.lte.auth_opc = hexToBase64(authOpc);
-        } else {
-          setError('auth_opc is not a valid hex');
-          return;
-        }
-      }
-
-      if (authKey !== '') {
-        if (isValidHex(authKey)) {
-          subscriberState.lte.auth_key = hexToBase64(authKey);
-        } else {
-          setError('auth_key is not a valid hex');
-          return;
-        }
-      }
-      const {config: _, ...mutableSubscriber} = {...subscriberState};
-      const staticIps = {};
-      subscriberStaticIPRows.forEach(
-        apn => (staticIps[apn.apnName] = apn.staticIp),
-      );
-      await ctx.setState?.(subscriberState.id, {
-        ...mutableSubscriber,
-        static_ips: staticIps,
-      });
-      enqueueSnackbar('Subscriber saved successfully', {
-        variant: 'success',
-      });
-    } catch (e) {
-      const errMsg = e.response?.data?.message ?? e.message;
-      setError('error saving ' + subscriberState.id + ' : ' + errMsg);
-      return;
-    }
-    props.onClose();
+  const subscriberProps = {
+    upload,
+    setUpload,
+    subscribers,
+    setSubscribers,
+    addError,
+    setAddError,
+    error: props.error,
+    onClose: props.onClose,
+    onSave: props.onSave,
+    rowAdd,
+    setRowAdd,
+    subscriberAction: props.subscriberAction,
   };
 
   return (
-    <Dialog
-      classes={{paper: classes.dialog}}
-      data-testid="editDialog"
-      open={props.open}
-      fullWidth={true}
-      maxWidth="sm">
-      <DialogTitle label={'Edit Subscriber Settings'} onClose={onClose} />
-      <Tabs
-        value={tabPos}
-        onChange={(_, v) => setTabPos(v)}
-        indicatorColor="primary"
-        className={classes.tabBar}>
-        <Tab
-          key="subscriber"
-          data-testid="subscriberTab"
-          label={SUBSCRIBER_TITLE}
-        />
-        ;
-        <Tab
-          key="trafficPolicy"
-          data-testid="trafficPolicyTab"
-          label={TRAFFIC_TITLE}
-        />
-        <Tab
-          key="apnStaticIps"
-          data-testid="staticIpsTab"
-          label={STATIC_IPS_TITLE}
-        />
-        ;
-      </Tabs>
-      <DialogContent>
-        <List>
-          {error !== '' && (
-            <AltFormField disableGutters label={''}>
-              <FormLabel data-testid="configEditError" error>
-                {error}
-              </FormLabel>
-            </AltFormField>
-          )}
-
-          {tabPos === 0 && (
-            <div>
-              <EditSubscriberDetails {...subscriberProps} />
-            </div>
-          )}
-          {tabPos === 1 && <EditSubscriberTrafficPolicy {...subscriberProps} />}
-          {tabPos === 2 && (
-            <div>
-              <EditSubscriberApnStaticIps {...subscriberProps} />
-            </div>
-          )}
-        </List>
-      </DialogContent>
-      <DialogActions>
-        <Button onClick={props.onClose} skin="regular">
-          {'Close'}
-        </Button>
-        <Button
-          data-testid={`${props.editProps?.editTable || ''}-saveButton`}
-          variant="contained"
-          color="primary"
-          onClick={onSave}>
-          {'Save'}
-        </Button>
-      </DialogActions>
-    </Dialog>
+    <>
+      {!(upload || props.subscriberAction === 'edit') ? (
+        <SubscriberDetailsTable {...subscriberProps} />
+      ) : (
+        <SubscriberDetailsUpload {...subscriberProps} />
+      )}
+    </>
   );
 }
 
-type AddSubscribersProps = {
+export type SubscribersDialogDetailProps = {
   // Subscribers to add, edit or delete
   setSubscribers: (Array<SubscriberInfo>) => void,
   subscribers: Array<SubscriberInfo>,
@@ -566,130 +182,16 @@ type AddSubscribersProps = {
   upload: boolean,
   onClose: () => void,
   // Add, edit or delete subscribers
-  onSave?: (Array<SubscriberInfo>, selectedSubscribers?: Array<string>) => void,
+  onSave: (Array<SubscriberInfo>, selectedSubscribers?: Array<string>) => void,
   error?: string,
   // Row added with the Add New Row button
   rowAdd: boolean,
   setRowAdd: boolean => void,
-  // Delete subscribers dialog
+  // Delete, Edit or Add subscriber
   subscriberAction: SubscriberActionType,
 };
 
-function SubscriberDetailsUpload(props: AddSubscribersProps) {
-  const {
-    setSubscribers,
-    setAddError,
-    setUpload,
-    upload,
-    subscribers,
-    subscriberAction,
-  } = props;
-  const classes = useStyles();
-  const enqueueSnackbar = useEnqueueSnackbar();
-  const [fileName, setFileName] = useState('');
-  const DropzoneText = () => (
-    <div>
-      Drag and drop or <Link>browse files</Link>
-    </div>
-  );
-
-  return (
-    <>
-      <DialogContent classes={{root: classes.uploadDialog}}>
-        <CardTitleRow label={'Upload CSV'} />
-        <Grid container>
-          <Grid item xs={12}>
-            <Alert severity="warning">
-              This will replace the subscribers you entered on the previous
-              page.
-            </Alert>
-          </Grid>
-          {!fileName ? (
-            <Grid item xs={12}>
-              <DropzoneArea
-                dropzoneText={<DropzoneText />}
-                useChipsForPreview
-                showPreviewsInDropzone={false}
-                filesLimit={1}
-                showAlerts={false}
-                onChange={async files => {
-                  if (files.length) {
-                    try {
-                      const newSubscribers = await parseSubscriberFile(
-                        files[0],
-                      );
-                      if (newSubscribers) {
-                        setSubscribers([...newSubscribers]);
-                        const errors = validateSubscribers(
-                          newSubscribers,
-                          subscriberAction === 'delete',
-                        );
-                        setFileName(files[0].name);
-                        if (!(subscriberAction === 'delete')) {
-                          setUpload(false);
-                          setAddError(errors);
-                        }
-                      }
-                    } catch (e) {
-                      enqueueSnackbar(e, {
-                        variant: 'error',
-                      });
-                    }
-                  }
-                }}
-              />
-              <Text variant="body2" className={classes.uploadInstructions}>
-                {`Accepted file type: .csv (<10 MB).  ${
-                  subscriberAction === 'delete'
-                    ? DELETE_INSTRUCTIONS
-                    : subscriberAction === 'add'
-                    ? ADD_INSTRUCTIONS
-                    : EDIT_INSTRUCTIONS
-                }`}
-                <Link href={UPLOAD_DOC_LINK}>documentation</Link>
-              </Text>
-            </Grid>
-          ) : (
-            <Grid item xs={12}>
-              <Alert severity="success">{`${fileName} is uploaded`}</Alert>
-            </Grid>
-          )}
-        </Grid>
-      </DialogContent>
-      <DialogActions>
-        <Grid container justify="space-between">
-          <Grid item>
-            {upload && (
-              <Button
-                onClick={() => {
-                  setUpload(false);
-                  if (subscriberAction === 'delete' && subscribers.length > 0) {
-                    setSubscribers([]);
-                  }
-                }}>
-                Back
-              </Button>
-            )}
-          </Grid>
-          <Grid item>
-            <Button onClick={props.onClose}> Cancel </Button>
-            <Button
-              data-testid="saveSubscriber"
-              onClick={() => {
-                props.onSave?.(subscribers);
-              }}>
-              {subscriberAction === 'delete' ?? false
-                ? 'Delete Subcribers'
-                : 'Save and Add Subscribers'}
-            </Button>
-          </Grid>
-        </Grid>
-      </DialogActions>
-    </>
-  );
-}
-
-function SubscriberDetailsTable(props: AddSubscribersProps) {
+function SubscriberDetailsTable(props: SubscribersDialogDetailProps) {
   const {
     setSubscribers,
     setAddError,
@@ -717,7 +219,7 @@ function SubscriberDetailsTable(props: AddSubscribersProps) {
   const tableActions = {
     onRowUpdate: (newData, oldData) => {
       return new Promise((resolve, reject) => {
-        const err = validateSubscribers([newData]);
+        const err = validateSubscribers([newData], subscriberAction);
         setAddError(err);
         if (err.length > 0) {
           return reject();
@@ -744,7 +246,6 @@ function SubscriberDetailsTable(props: AddSubscribersProps) {
   const {match} = useRouter();
   const [maxPageRowCount, setMaxPageRowCount] = useState(0);
   const [tokenList, setTokenList] = useState(['']);
-
   const networkId: string = nullthrows(match.params.networkId);
   const subscriberMetrics = ctx.metrics;
   const getSubscribers = (query: ActionQuery) =>
@@ -1003,8 +504,9 @@ function SubscriberDetailsTable(props: AddSubscribersProps) {
                   {
                     title: '',
                     field: '',
-                    width: '5%',
+                    width: '70px',
                     editable: 'never',
+                    align: 'center',
                     render: rowData => (
                       <Text variant="subtitle3">
                         {rowData.tableData?.id + 1 || ''}
@@ -1059,7 +561,10 @@ function SubscriberDetailsTable(props: AddSubscribersProps) {
                   onRowAdd: newData => {
                     setRowAdd(true);
                     return new Promise((resolve, reject) => {
-                      const err = validateSubscribers([newData]);
+                      const err = validateSubscribers(
+                        [newData],
+                        subscriberAction,
+                      );
                       setAddError(err);
                       if (err.length > 0) {
                         return reject();
@@ -1113,7 +618,7 @@ function SubscriberDetailsTable(props: AddSubscribersProps) {
               color="primary"
               data-testid="saveSubscriber"
               onClick={() => {
-                const err = validateSubscribers(subscribers);
+                const err = validateSubscribers(subscribers, subscriberAction);
                 setAddError(err);
                 if (!err.length) {
                   props.onSave?.(subscribers, selectedSubscribers);
@@ -1128,232 +633,5 @@ function SubscriberDetailsTable(props: AddSubscribersProps) {
         </Grid>
       </DialogActions>
     </>
-  );
-}
-
-/**
- * Dialog content used to Add/Delete/Update subscribers
- * Displays upload subscriber dropzone or subscriber table
- */
-function SubscriberDetailsDialogContent(props: ActionDialogProps) {
-  const [addError, setAddError] = useState([]);
-  const [subscribers, setSubscribers] = useState<Array<SubscriberInfo>>([]);
-  const [upload, setUpload] = useState(false);
-  const [rowAdd, setRowAdd] = useState(false);
-
-  const addSubscriberProps = {
-    upload,
-    setUpload,
-    subscribers,
-    setSubscribers,
-    addError,
-    setAddError,
-    error: props.error,
-    onClose: props.onClose,
-    onSave: props.onSave,
-    rowAdd,
-    setRowAdd,
-    subscriberAction: props.subscriberAction,
-  };
-
-  return (
-    <>
-      {!upload ? (
-        <SubscriberDetailsTable {...addSubscriberProps} />
-      ) : (
-        <SubscriberDetailsUpload {...addSubscriberProps} />
-      )}
-    </>
-  );
-}
-
-export function validateSubscriberInfo(
-  info: SubscriberInfo,
-  subscribers: {[string]: subscriber},
-  edit?: boolean,
-) {
-  if (!info.imsi.match(/^(IMSI\d{10,15})$/)) {
-    return "imsi invalid, should match '^(IMSId{10,15})$'";
-  }
-  if (info.imsi in subscribers && !(edit ?? false)) {
-    return 'imsi already exists';
-  }
-  if (info.authKey && !isValidHex(info.authKey)) {
-    return 'auth key is not a valid hex';
-  }
-  if (info.authOpc && !isValidHex(info.authOpc)) {
-    return 'auth opc is not a valid hex';
-  }
-  return '';
-}
-
-type SubscriberError = $Values<typeof SUBSCRIBER_ADD_ERRORS>;
-
-/**
- * Checks subscriber fields format
- *
- * @param {Array<SubscriberInfo>} subscribers Array of subcribers to validate
- * @returns {Array<string>} Returns array of errors
- */
-export function validateSubscribers(
-  subscribers: Array<SubscriberInfo>,
-  isDelete?: boolean,
-) {
-  const errors: {
-    [error: SubscriberError]: Array<number>,
-  } = {};
-  const imsiList = [];
-
-  Object.keys(SUBSCRIBER_ADD_ERRORS).map(error => {
-    const subscriberError = SUBSCRIBER_ADD_ERRORS[error];
-    errors[subscriberError] = [];
-  });
-  subscribers.forEach((info, i) => {
-    if (!info.authKey && !(isDelete ?? false)) {
-      errors[SUBSCRIBER_ADD_ERRORS['REQUIRED_AUTH_KEY']].push(i + 1);
-    }
-    if (!info?.imsi?.match(/^(IMSI\d{10,15})$/)) {
-      errors[SUBSCRIBER_ADD_ERRORS['INVALID_IMSI']].push(i + 1);
-    }
-    if (info.authKey && !isValidHex(info.authKey)) {
-      errors[SUBSCRIBER_ADD_ERRORS['INVALID_AUTH_KEY']].push(i + 1);
-    }
-    if (info.authOpc && !isValidHex(info.authOpc)) {
-      errors[SUBSCRIBER_ADD_ERRORS['INVALID_AUTH_OPC']].push(i + 1);
-    }
-    if (!info.dataPlan && !(isDelete ?? false)) {
-      errors[SUBSCRIBER_ADD_ERRORS['REQUIRED_SUB_PROFILE']].push(i + 1);
-    }
-    if (imsiList.includes(info.imsi) && !(isDelete ?? false)) {
-      errors[SUBSCRIBER_ADD_ERRORS['DUPLICATE_IMSI']].push(i + 1);
-    } else {
-      imsiList.push(info.imsi);
-    }
-  });
-
-  const errorList: Array<string> = Object.keys(SUBSCRIBER_ADD_ERRORS)
-    .map(error => SUBSCRIBER_ADD_ERRORS[error])
-    .reduce((res, errorMessage) => {
-      if (errors[errorMessage].length > 0) {
-        res.push(
-          `${errorMessage} : Row ${errors[errorMessage].sort().join(', ')}`,
-        );
-      }
-      return res;
-    }, []);
-
-  return errorList;
-}
-
-type subscriberStaticIpsRowType = {
-  apnName: string,
-  staticIp: string,
-};
-
-type subscriberForbiddenNetworkTypes = {
-  nwTypes: string,
-};
-
-function EditSubscriberDetails(props: EditSubscriberProps) {
-  const classes = useStyles();
-  return (
-    <div>
-      <List>
-        <AltFormField label={'Subscriber Name'}>
-          <OutlinedInput
-            data-testid="name"
-            className={classes.input}
-            placeholder="Enter Name"
-            fullWidth={true}
-            value={props.subscriberState.name}
-            onChange={({target}) =>
-              props.onSubscriberChange('name', target.value)
-            }
-          />
-        </AltFormField>
-        <AltFormField label={'Service State'}>
-          <TypedSelect
-            className={classes.input}
-            input={<OutlinedInput />}
-            value={props.subscriberState.lte.state}
-            items={{
-              ACTIVE: 'Active',
-              INACTIVE: 'Inactive',
-            }}
-            onChange={value => {
-              props.onSubscriberChange('lte', {
-                ...props.subscriberState.lte,
-                state: value,
-              });
-            }}
-          />
-        </AltFormField>
-        <AltFormField label={'Data Plan'}>
-          <TypedSelect
-            className={classes.input}
-            input={<OutlinedInput />}
-            value={props.subscriberState.lte.sub_profile}
-            items={props.subProfiles}
-            onChange={value => {
-              props.onSubscriberChange('lte', {
-                ...props.subscriberState.lte,
-                sub_profile: value,
-              });
-            }}
-          />
-        </AltFormField>
-        <AltFormField label={'Forbidden Network Types'}>
-          <FormControl className={classes.input}>
-            <Select
-              multiple
-              value={props.subscriberState.forbidden_network_types ?? []}
-              onChange={({target}) => {
-                props.onSubscriberChange(
-                  'forbidden_network_types',
-                  target.value,
-                );
-              }}
-              renderValue={selected => selected.join(', ')}
-              input={<OutlinedInput />}>
-              {forbiddenNetworkTypes.map((k: string, idx: number) => (
-                <MenuItem key={idx} value={k}>
-                  <Checkbox
-                    checked={
-                      props.subscriberState.forbidden_network_types != null
-                        ? props.subscriberState.forbidden_network_types.indexOf(
-                            k,
-                          ) > -1
-                        : false
-                    }
-                  />
-                  <ListItemText primary={k} />
-                </MenuItem>
-              ))}
-            </Select>
-          </FormControl>
-        </AltFormField>
-
-        <AltFormField label={'Auth Key'}>
-          <PasswordInput
-            data-testid="authKey"
-            className={classes.input}
-            placeholder="Eg. 8baf473f2f8fd09487cccbd7097c6862"
-            value={props.authKey}
-            error={props.authKey && !isValidHex(props.authKey) ? true : false}
-            onChange={v => props.setAuthKey(v)}
-          />
-        </AltFormField>
-        <AltFormField label={'Auth OPC'}>
-          <PasswordInput
-            data-testid="authOPC"
-            value={props.authOpc}
-            placeholder="Eg. 8e27b6af0e692e750f32667a3b14605d"
-            className={classes.input}
-            error={props.authOpc && !isValidHex(props.authOpc) ? true : false}
-            onChange={v => props.setAuthOpc(v)}
-          />
-        </AltFormField>
-      </List>
-    </div>
   );
 }
